Fixed generation of pot file (now uses msghack instead of msgcat).
For reasons unknown to me msgcat doesn't work with files generated with
xgettext for this project. It seems to be working in AqBanking but not here
even though AqBanking uses the same approach...

The problem is that xgettext doesn't seem to set a correct charset header
(it only writes charset=CHARSET on which msgcat chokes).

msghack hoever accepts the input so we use that for now.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1710 70169cfe-8b10-0410-8925-dcb4b91034d8