Added function GWEN_Gui_GetSyncIo().
This allows for application-defined connection handling because the
application can create the GWEN_SYNCIO necessary to handle a given URL.
For HTTP requests the default implementation creates a socket layer,
a buffered io layer and a http layer. For HTTPS a tls layer is inserted
between socket and buffered io.

Applications can overwrite this function and create PROXY layers instead.
That makes it possible to use PROXY handling provided by desktop environments
like KDE and GNOME.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1729 70169cfe-8b10-0410-8925-dcb4b91034d8