More improvements on GTK2 implementation.
We need to let parent widgets do the adding of child widgets because
GTK2 seems to be more parent-centered than other toolkits.

In GTK2 one has to know the type of the parent widget to decide how to
add a child (e.g. either call gtk_box_pack_* or gtk_container_add, and
with GtkTable there is even another function for that purpose).

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1775 70169cfe-8b10-0410-8925-dcb4b91034d8