Fixed a bug with some https servers (especially ing-diba).
There is a problem if the bank doesn't send a "Content-Length:". In that case
we have to read until the bank closes the connection.
However, GnuTLS doesn't return a 0 upon recv but rather returns a "-9" error
code if the connection has been closed by the peer :-/
So we have to deal with it like previous versions of gwen did: Ignore the
error code if at least a few bytes have been received.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1790 70169cfe-8b10-0410-8925-dcb4b91034d8