MacOSX fixes: Improved relocatability.
Use _NSGetExecutablePath() instead of CFBundleCopyResourcesDirectoryURL()
because the latter points to the Resource folder within a bundle which isn't
what we expect from the function GWEN_Directory_GetPrefixDirectory().

However, this might only be available since OSX 10.5. But thats the minimum
supported OSX version anyway.

Changed the default paths for MacOSX in local-install-mode. This allows to
include AqBanking with application bundles.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1795 70169cfe-8b10-0410-8925-dcb4b91034d8