Fix off-by-one error when checking buffer size.
Pointed out by Michael Haller. The argument maxLen should better be
called bufferSize because it designates the allowed number of bytes in
the buffer which includes the trailing zero. The maximum number of input
characters is one less. The documentation in gui.h explains this correctly,
but the argument would still be more understandable when called bufferSize.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2021 70169cfe-8b10-0410-8925-dcb4b91034d8