Started working on a simple archiver.
This will be used by future versions of AqFinance as a replacement for
tar. tar is an excellent tool, but libtar is very hard to compile on
non-POSIX systems like Windows (at least I couldn't get it to work on WIN32
platforms). So for AqFinance to support automatic update management I need
a simple replacement which only allows safely adding multiple files and
folders to an archive file and to extract them later. No funny stuff.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2092 70169cfe-8b10-0410-8925-dcb4b91034d8