First step in adapting to latest API changes in libgcrypt.
Current versions of libgcrypt mark the alternate interface to public keys
as deprecated, so we need to move our code to no longer use that API.

Fortunately, this is quite easy since gwen only used that to store public
keys and not for cryptographic functions. As soon as this new module has
been tested more extensively it will replace the old keyRsa module
transparently (e.g. the suffix "2" will disappear).

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2137 70169cfe-8b10-0410-8925-dcb4b91034d8