Removed password storage code from console gui code.
This is now completely implemented in the base GWEN_GUI class.

Added flag "GWEN_CGUI_FLAGS_PERMPASSWORDS", this is used to force storing
passwords in the permanent password store, if there is one. The reason for
this flag is that in console code we can't ask the user conveniently
whether the just entered password is allowed to be stored in the password
storage (we would have to ask the user after asking for the password
itself).
In GUI code the dialog might just have an additional checkbox
("Store permanently"), but we don't have that on the console.

Added function "GWEN_PasswordStore_ClearStoragePasswd". This is used
indirectly by AqBanking after every bank contact session.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2171 70169cfe-8b10-0410-8925-dcb4b91034d8