Typemaker2: Improved copy function
Added copyFlags to members of a type and use them when creating the
copy function.

This can be used to distinguish between copy and dup operations where
dup normally creates a deep copy of the source object while copy only
copies some parts of an object.

However, this is now completely up to the designer of the type file.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2227 70169cfe-8b10-0410-8925-dcb4b91034d8