Add missing linker flags for libgpg-error.so.0
This fixes a warning found by the Debian build log checks. More detailed
information can be found at [1]. The build logs contains a line like the
following:

dpkg-shlibdeps: warning: symbol gpg_strerror used by libgwenhywfar.so.60.6.0 found in none of the libraries

In this case the reason is that libgwenhywfar.so.60.6.0 is not linked against
libgpg-error.so.0, even though gpg_strerror() from that library is being used
in the code.

Apparently this didn't cause any trouble (up to now), because libgpg-error.so.0
is loaded by the dynamic linker due to an indirect linker dependency on it
(e.g. because one of the linked libraries is itself linked against it).

However, better be safe than sorry, so let's explicitly link all libraries that
we use symbols from.

[1] https://buildd.debian.org/~brlink/bytag/W-shlibs-symbol-not-found.html


git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2232 70169cfe-8b10-0410-8925-dcb4b91034d8