TLS: Fixed another problem.
The string
 "SECURE256:SECURE128:-ARCFOUR-128:-AES-128-CBC:-CAMELLIA-128-CBC:-3DES-CBC"
makes it impossible to connect to postbank.de. For whatever reason Postbank
tries to enforce ARCFOUR-128 when given the opportunity to use a 128 bit
algo, despite the fact that our list starts with SECURE256...

Removing "SECURE128" leaves Postbank only the option to use a 256 bit algo,
and that works.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2278 70169cfe-8b10-0410-8925-dcb4b91034d8