Make GWEN_StringList_Sort() compliant with the documented behaviour (Elias Oltmanns)
Contrary to the documentation in stringlist.h, GWEN_StringList_Sort()
used to sort in ascending order when the argument ascending was *not*
set, and vice versa. This bug has been (partially) masked by the fact,
that sorting was broken, anyway, due to incorrect book keeping with
regard to the number of list elements and inappropriate integer
sorting functions.

This change will likely require adjustments to code in other
applications and libraries, e.g. AqBanking. Unfortunately, a simple
0/1 swap of the ascending parameter will not do because the bug
evidently did not occur consistently.

git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@2296 70169cfe-8b10-0410-8925-dcb4b91034d8