Added class GWEN_PARAM.
This is used at least by AqRadBase, and probably by my other projects
as well. It can be used to provide descriptions for settings and a framework
to manage those settings.

AqRadBase will use it to manage the settings of ADCM_CONVERTER and
ADCM_OPERATOR classes. Those settings can then be provided in a generic
manner.