MsgEngine: Properly parse messages containing optional fields
Consider the case that a message definition contains two elements (or
groups) with varying multiplicty at the end. If the latter has a
value (or several values, for that matter), this can only be
unambiguously expressed if the former element has a specified maximum
number of repeats and exactly that many delimiters have been provided.
Make the parser rely on (and honour) that requirement.