Extended GWEN_Gui_GetPasword().
The pasword entry function is used by homebanking apps/libraries for
input of passwords for keyfiles, for pin input and for tan input.
FinTS in Germany uses several methods to get a tan; some just ask
for transaction number (TAN) from a previously mailed list.  Others
ask you to hold a special tan generator device to the monitor to let
it read a special flicker graphic. Some show an image to the user from
which he must derive the tan in some way. Therefore the password entry
function needed to be extended to allow for a more generic way to ask
the user for a password. Previously, some specific data was included in
the text parameter, from which the GUI implementation needed to extract
that data (i.e. "$OBEGIN" and "$OEND" for optical data.
This should become much easier now.