Typemaker2: Allow for types to have special includes to be used in structs.
structs are defined by a list of members. Each member is of a specific
type, and each of those types (e.g. "uint8_t_array") can now have its own
"includes" section, which hold functions to be added for members of those
types.

E.g. "uint8_t_array" now includes member functions, which allow setting
items of an array at given indices. So when using "uint8_t_array" in a
struct which contains a member of that type with the name "testArray" then
the generated struct will have (upon others) these additional functions:

- uint8_t MyStruct_GetTestArrayAt(const MYSTRUCT *myStruct, int idx);
- int MyStruct_SetTestArrayAt(MYSTRUCT *myStruct, int idx, uint8_t v);
- void MyStruct_PresetTestArray(MYSTRUCT *myStruct, uint8_t v);

Added variable "nameWithCapital" which contains a member name which starts
with a capital.