pkg-config: Fix pkgincludedir variable
This variable got out of sync with where the header files are actually
installed to. Apparently nobody noticed this so far because it is barely used.

While at it, also use @gwenhywfar_headerdir@ instead of hardcoded path for the
includedir variable. This avoids includedir to get out of sync in the future.