Typemaker2: Add support for member specific variables.
When resolving <includes> or function calls we already had member
variables like "$(maxlen)" and others. However, especially in <include>
section we sometimes need information about members which is not available.
For this purpose we now create those variables for every member of a struct
by prefixing it with the member name and ".".

So for example in a struct which contains a member called "testIntegerArray"
we can use the variable "$(testIntegerArray.maxlen)" inside an <include>
section to access the maximum length of that member.