Add version info in gui related shared libraries
Distributions manage package dependencies based on the
filename of shared libraries (at least with openSUSE,
which I know of), which doesn't work in the case of
the gwengui-xxx libraries, because they don't contain
any or always the same version information (=0).

This can lead to the installation of an unsuitable
version if two sources are configured on one
installation containing different versions of
gwenhywfar.

To avoid this, the libraries mentioned above will
receive appropriate version information with this
commit.