SIO: Added call to gnutls_server_name_set(), should enable SNI.
Setting the server name should enable SNI extension in GnuTLS according
to the GnuTLS APIDOC.

If this extension is not available (or hostnam not set) then the resulting
error is ignored for now. The result of this will be a warning from GnuTLS
that the received certificate was not issued for the connected host.