GWEN_IdList64: Make argument to GWEN_IdList64_Iterator_new() const.
From now on GWEN_IdList64_Iterator_new() will no longer attach to the
original list which makes it possible to change the argument to const.
This also means that the iterator class can be used even in multithreaded
environments.

This represents an API change, however, no code so far relied on the fact
that GWEN_IdList64_Iterator_new() used to call GWEN_IdList64_Attach().