OS: More work on threads.
Added functions to set flags. Those are used when creating a thread.
The only flag currently available allows for creating undetached threads
(currently only on POSIX systems, on Windows the resources must always be
released with CloseHandle()).

Need to find a way to portably find out whether a thread is still running...