CryptToken: No longer use GWEN_Gui_ShowBox() and GWEN_Gui_HideBox().
Those GUI functions need the GUI to run in the background which doesn't work
well in most apps where the program code runs inside the same thread as the
GUI code.
Therefore sometimes dialog boxes aren't correctly refreshed or shown on
some platforms.

A message in the log window seems to be the best and simplest solution to
the surprising complexity of the problem.