GwenBuild: Build "tm2builder" targets first if none specified.
Source files generated from typemaker2 type definition files in most cases
need to be generated before anything else. Therefore we prioritize those
builders if no specific builder is requested on the command line.

If a specific builder is given on the command line then only builders of that
given type are run, so in that case the behaviour is not changed.