CSV: Cleanup, allow for subfields.
The CSV importer can now also handle subfields and parse them accordingly.
This is usefull if a CSV field contains more structured data.

In such a case the column spec in the configuratin file must contain a
DB_Group instead of a variable like in this example:

  columns {
    1="date"
    2="valutadate"
    3="type"
    4="num"
    5 {
      delimiter="/"
      1="wkn"
      2="isin"
    }
    6="totalAmountValue"
    7="totalAmountCurrency"
    8="priceValue"
    9="priceCurrency"
  } # columns

Here the 5th field contains "WKN" and "ISIN" which are separated by "/".