typemaker2: Fixed a bug in _buildMemberInlines()
We need to make sure member type is not the same type as the struct type,
otherwise the inlines will be generated multiple times (once for the main
struct and once for every member type which is of the same type).

The idea of member inlines was to automatically add functions like
addFlags() and subFlags() for uint32_t members which are used as flags.