gwbuild: Use new redirection code of GWEN_Process module.
This finally creates consistent output for gwbuild. Previously we didn't
redirect output of the created compile/link/typemaker2 processes so they
wrote to the console whenever they felt like it, sometimes scrambling the
console output.

Now stdout and stderr are redirected to temporary logfiles which are then
printed after a command finished. This makes also warnings and other
messages available to the console output.