plugin: Combine plugin descriptions from disk and added plugins.
Gwenhywfar is able to load plugins. However, this can be restricted when
building gwen in a fully static environment which don't allow for dynamic
loading of plugins.

In such a case plugin providers can add just create a plugin and add it to
Gwen's internal list.

However, an application querying the list of available plugins of a type
previously had to rely on there being a plugin description on disk.

This is no longer the case. Applications or libraries can add plugins to
existing plugin managers and also provide a plugin description for it.