Port deprecated QFlags constructor
When consumers want to build without deprecated Qt public API using for
example `-DQT_DISABLE_DEPRECATED_BEFORE=0x50f02` as a compile
definition, this line breaks the build. Adding this is currently in the
works for KMyMoney.

The fix is to just use the default constructor or an initializer list
instead. Because the initializer list is available in Qt 5.4 and the
default constructor in 5.15, the initializer list syntax is used.
See See https://doc.qt.io/qt-5/qflags-obsolete.html.

Signed-off-by: Alexander Lohnau <alexander.lohnau@gmx.de>
Signed-off-by: Thomas Baumgart <thb@net-bembel.de>