msgio: introduced callbacks around writing packets
The AqHome project's wire protocol uses a BUSY line to mark/detect whether
the bus is currently in use.

Before sending a packet over the wire the BUSY line needs to be checked
and if free pulled down to mark the bus as BUSY. After sending the packet
the BUSY line is raised. This is now possible using the new callbacks.