Fix cross compilation
libgwenhywfar failed to cross build from source, because it fails running
mklistdoc with an "Exec format error". This happens when attempting to
run a tool that is built for the host architecture. mklistdoc really
needs to be built for the host architecture, because it is installed
into gwenhywfar-tools. On the flip side that means that we can just run
the mklistdoc from a system gwenhywfar-tools. Note that this change only
affects cross compilation. In native builds, the built mklistdoc will
continue to be used.

These changes were initially reported by Helmut Grohne <helmut@subdivi.de> via
the Debian bug tracker (see https://bugs.debian.org/1051173).