Fix br_init_lib to work with modern compilers
More modern compiler place the string used to locate the filename of the
library in a different (non-executable) segment. Therefore, the routine
to find the filename failed.  This problem only shows one configures
with --enable-binreloc which is used when packaging it into an AppImage
format.

Using a code symbol of the library as reference solves the problem.